Create separate file for data table types.
