+++ graylog2-web-interface/src/components/common/ConfigurableDataTable/ConfigurableDataTable.tsx
@@ -25,27 +25,7 @@ import TableHead from 'components/common/ConfigurableDataTable/TableHead';
 import TableBody from 'components/common/ConfigurableDataTable/TableBody';
 import useCurrentUser from 'hooks/useCurrentUser';
 
-export type Attribute = {
-  id: string,
-  title: string,
-  type?: boolean,
-};
-
-export type CustomHeaders = {
-  [key: string]: {
-    renderHead: (attribute: Attribute) => React.ReactNode,
-    textAlign: string,
-  }
-}
-
-export type CustomCells<ListItem extends { id: string }> = {
-  [attributeId: string]: {
-    renderCell: (listItem: ListItem, attribute: Attribute) => React.ReactNode,
-    textAlign?: string,
-    width?: string,
-    maxWidth?: string,
-  }
-}
+import type { CustomCells, CustomHeaders, Attribute } from './types';
 
 const ScrollContainer = styled.div`
   overflow-x: auto;
